Simplify lockscreen setup
So it turns out that xtrlock-pam does exactly what I want: it shows the
contents of the root window, also known as the desktop background.
Further, home-manager already has a $screen-locker service that locks on
dpms blank, suspend, and through loginctl and xdg-screensaver commands.
This all sounds nice to me.