Edit: Exit on ESC in skim
It turns out that the output substitution was causing skim's
exit code to be ignored, and the if would use the exit code
of readarray instead. This leads to the undesirable behavior
that edit will open with no file selected when ESC is
pressed.

This commit fixes that by using the `-e` flag for bash to
exit on all errors, and dropping the if statement